requirements.txt
python-logging-loki
requests
python-dotenv